Complete basic unit tests for ImageGenerator.tsx
Change the "current" field in IQuery to a "prompt" model, with input and output like the "history" property array in IQuery. This makes swapping out the current input for an item in the query history a lot easier.

Also, I split the IQuery interface into ITextQuery and IImageQuery. This involved some other refactoring.