Add ArticleList.tsx & ArticleListItem.tsx components
The profile page will utilize these components to showcase previously written articles. Although the components currently have no implementation, they do have corresponding Jest test files. Additionally, I have added a conditional statement in the articles API that checks for the absence of parameters. If none are present, it is assumed to be a request for all articles belonging to a user and will return accordingly. This will be used in the aforementioned components or possibly obtained directly from the user model.