Complete ArticleList.tsx & ArticleListItem.tsx
Please note that this change only includes functional updates and does not encompass any styling modifications.

While there is still room for improvement, the tests for the delete function in the ArticleListItem.test.tsx file has some limitations. Specifically, the delete test is unable to verify the update of the SWR state after a mutation in the ArticleList.tsx component. However, this could (and probably should) be tested in ArticleList.test.tsx.