Rearrange page structure
Move the account page into the admin section so all admin pages can share the same dashboard-styled "layout".

Rename /article page directory to /articles and add a page at the root of the /articles directory. This allows me to create an "articles" pages that will list all the user's articles.