Transform "/subscrition/page.tsx" to "/subscription/[id]/page.tsx"
Make subscription a dynamic route with the "[id]" variable being the price ID for that subscription as found on Stripe. If in the future more pricing tiers are introduced, we need only create them on stripe and add their price ID to related components. The relevant PaymentIntent will be created.