Adding emailing functionality via "@lib/email"
This change introduces the ability to send emails from the application using the Nodemailer library. The code is housed in the email folder, which contains the services folder that holds the EmailService class responsible for sending emails.

The class takes in the recipient's email address, subject, HTML content, and CSS styles as arguments and uses the Nodemailer library to send the email. The configuration for the SMTP server is specified in the EmailService class constructor.

To use the email service, simply import the EmailService class and create an instance of it, passing in the necessary arguments. The class can then be used to send emails as needed.

This change adds the capability to send stylish emails from the application and enhances the overall functionality of the project.