Update user.model.ts with Stripe subscription fields
The 3 fields are used to store information about the customer's Stripe account.

stripeId: A unique identifier generated by the Stripe API that represents a customer in the system.

subscriptionId: A unique identifier generated by the Stripe API that represents a subscription in the Stripe system.

subscriptionStatus: A string that represents the current status of the customer's subscription on Stripe. It can be one of the following values: active, past_due, or cancelled.

Having these fields in the user model allows you to easily store and manage customer information and their subscription status within your own application database.